% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{add_enthalpy_label}
\alias{add_enthalpy_label}
\title{Add Enthalpy Label}
\usage{
add_enthalpy_label(plot, line_data, h)
}
\arguments{
\item{plot}{ggplot object}

\item{line_data}{Line data for enthalpy}

\item{h}{Enthalpy value}
}
\value{
ggplot object with label added
}
\description{
Internal function to add enthalpy labels to plot
}
\keyword{internal}
