% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{add_occurrence_scale}
\alias{add_occurrence_scale}
\title{Add Occurrence Color Scale}
\usage{
add_occurrence_scale(color_palette, color_limit)
}
\arguments{
\item{color_palette}{Color palette name}

\item{color_limit}{Maximum value for color scale}
}
\value{
ggplot scale object
}
\description{
Internal function to add occurrence-based color scale
}
\keyword{internal}
