% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{add_psychrometric_lines}
\alias{add_psychrometric_lines}
\title{Add Psychrometric Chart Lines}
\usage{
add_psychrometric_lines(plot)
}
\arguments{
\item{plot}{ggplot object}
}
\value{
ggplot object with lines added
}
\description{
Add Psychrometric Chart Lines
}
\keyword{internal}
