% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{add_quantile_scale}
\alias{add_quantile_scale}
\title{Add Quantile Color Scale}
\usage{
add_quantile_scale(color_palette, n)
}
\arguments{
\item{color_palette}{Color palette name}

\item{n}{Number of colors needed}
}
\value{
ggplot scale object
}
\description{
Internal function to add quantile-based color scale
}
\keyword{internal}
