% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{apply_plot_theme}
\alias{apply_plot_theme}
\title{Apply Plot Theme}
\usage{
apply_plot_theme(plot)
}
\arguments{
\item{plot}{ggplot object}
}
\value{
ggplot object with theme applied
}
\description{
Apply Plot Theme
}
\keyword{internal}
