% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{calculate_global_color_limit}
\alias{calculate_global_color_limit}
\title{Calculate Global Color Limit}
\usage{
calculate_global_color_limit(data_df, interval, time_labels)
}
\arguments{
\item{data_df}{Data frame}

\item{interval}{Time interval}

\item{time_labels}{Time labels}
}
\value{
Global color limit
}
\description{
Internal function to calculate global color limit for animation
}
\keyword{internal}
