% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{calculate_quantiles}
\alias{calculate_quantiles}
\title{Calculate Quantiles for Data}
\usage{
calculate_quantiles(combinations_df, qvec)
}
\arguments{
\item{combinations_df}{Data frame with occurrence counts}

\item{qvec}{Quantile values}
}
\value{
Data frame with quantile information
}
\description{
Internal function to calculate quantiles for coloring
}
\keyword{internal}
