% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{cleanup_temp_files}
\alias{cleanup_temp_files}
\title{Clean Up Temporary Files}
\usage{
cleanup_temp_files(temp_files, name)
}
\arguments{
\item{temp_files}{Temporary files to remove}

\item{name}{Base name used for static plot}
}
\description{
Internal function to clean up temporary files
}
\keyword{internal}
