% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_animation_frames}
\alias{create_animation_frames}
\title{Create Animation Frames}
\usage{
create_animation_frames(
  data_df,
  title,
  color_palette,
  color_limit,
  quantiles,
  qvec,
  name,
  interval,
  time_labels
)
}
\arguments{
\item{data_df}{Data frame}

\item{title}{Plot title}

\item{color_palette}{Color palette}

\item{color_limit}{Color limit}

\item{quantiles}{Whether to use quantiles}

\item{qvec}{Quantile values}

\item{name}{Base name}

\item{interval}{Time interval}

\item{time_labels}{Time labels}
}
\value{
Vector of temporary file names
}
\description{
Internal function to create individual animation frames
}
\keyword{internal}
