% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_base_plot}
\alias{create_base_plot}
\title{Create Base Plot}
\usage{
create_base_plot(combinations_df, color_column, title)
}
\arguments{
\item{combinations_df}{Data frame with plot data}

\item{color_column}{Column to use for coloring}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Internal function to create the base ggplot
}
\keyword{internal}
