% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric_calculations.R
\name{create_line_data}
\alias{create_line_data}
\title{Create Line Data for Relative Humidity}
\usage{
create_line_data(Hr)
}
\arguments{
\item{Hr}{Relative humidity value}
}
\value{
Data frame with temperature and specific humidity coordinates
}
\description{
Create Line Data for Relative Humidity
}
