% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric_calculations.R
\name{create_line_data2}
\alias{create_line_data2}
\title{Create Line Data for Enthalpy}
\usage{
create_line_data2(h)
}
\arguments{
\item{h}{Enthalpy value}
}
\value{
Data frame with temperature and specific humidity coordinates
}
\description{
Create Line Data for Enthalpy
}
