% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_plot_title}
\alias{create_plot_title}
\title{Create Plot Title}
\usage{
create_plot_title(base_title, interval, time_label)
}
\arguments{
\item{base_title}{Base title}

\item{interval}{Time interval}

\item{time_label}{Current time label}
}
\value{
Formatted plot title
}
\description{
Internal function to create plot title for animation frames
}
\keyword{internal}
