% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_static_plot}
\alias{create_static_plot}
\title{Create Static Psychrometric Plot}
\usage{
create_static_plot(
  data_df,
  title,
  color_palette,
  color_limit = NULL,
  quantiles,
  qvec,
  name
)
}
\arguments{
\item{data_df}{Data frame with climate data}

\item{title}{Plot title}

\item{color_palette}{Color palette name}

\item{color_limit}{Maximum value for color scale}

\item{quantiles}{Whether to use quantile-based coloring}

\item{qvec}{Quantile values for coloring}

\item{name}{Base name for output file}
}
\value{
ggplot object
}
\description{
Create Static Psychrometric Plot
}
\keyword{internal}
