% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric_calculations.R
\name{dataframeplot}
\alias{dataframeplot}
\title{Create Data Frame for Plot Grid}
\usage{
dataframeplot()
}
\value{
Data frame with temperature and SH combinations
}
\description{
Create Data Frame for Plot Grid
}
