% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{extract_time_info}
\alias{extract_time_info}
\title{Extract Time Information}
\usage{
extract_time_info(data_df, interval)
}
\arguments{
\item{data_df}{Data frame}

\item{interval}{Time interval}
}
\value{
List with time units and labels
}
\description{
Internal function to extract time information for animation
}
\keyword{internal}
