% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{filter_by_date_range}
\alias{filter_by_date_range}
\title{Filter Data by Date Range}
\usage{
filter_by_date_range(data_df, date_range)
}
\arguments{
\item{data_df}{Data frame}

\item{date_range}{Date range vector}
}
\value{
Filtered data frame
}
\description{
Internal function to filter data by date range
}
\keyword{internal}
