% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{filter_by_year}
\alias{filter_by_year}
\title{Filter Data by Year}
\usage{
filter_by_year(data_df, year)
}
\arguments{
\item{data_df}{Data frame to filter}

\item{year}{Year to filter by}
}
\value{
Filtered data frame
}
\description{
Internal function to filter data by year
}
\keyword{internal}
