% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{filter_data_by_time}
\alias{filter_data_by_time}
\title{Filter Data by Time}
\usage{
filter_data_by_time(data_df, interval, time_label)
}
\arguments{
\item{data_df}{Data frame}

\item{interval}{Time interval}

\item{time_label}{Current time label}
}
\value{
Filtered data frame
}
\description{
Internal function to filter data by time interval
}
\keyword{internal}
