% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_color_palette}
\alias{get_color_palette}
\title{Get Color Palette}
\usage{
get_color_palette(color_palette = NULL)
}
\arguments{
\item{color_palette}{Name of color palette or custom color vector}
}
\value{
Color vector
}
\description{
Get Color Palette
}
\keyword{internal}
