% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_predefined_palette}
\alias{get_predefined_palette}
\title{Get Predefined Color Palette}
\usage{
get_predefined_palette(palette_name)
}
\arguments{
\item{palette_name}{Palette name}
}
\value{
Color vector
}
\description{
Internal function to get predefined color palettes
}
\keyword{internal}
