% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_predefined_quantile_palette}
\alias{get_predefined_quantile_palette}
\title{Get Predefined Quantile Palette}
\usage{
get_predefined_quantile_palette(palette_name, n)
}
\arguments{
\item{palette_name}{Palette name}

\item{n}{Number of colors}
}
\value{
Color vector
}
\description{
Internal function to get predefined palettes for quantiles
}
\keyword{internal}
