% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_quantile_palette}
\alias{get_quantile_palette}
\title{Get Color Palette for Quantiles}
\usage{
get_quantile_palette(palette_name = NULL, n = 4)
}
\arguments{
\item{palette_name}{Name of color palette or custom color vector}

\item{n}{Number of colors needed (based on quantiles)}
}
\value{
Color vector for quantiles of appropriate length
}
\description{
Get Color Palette for Quantiles
}
\keyword{internal}
