% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric_calculations.R
\name{maskgrid}
\alias{maskgrid}
\title{Create Grid Lines for Psychrometric Chart}
\usage{
maskgrid()
}
\value{
List containing vertical and horizontal grid line data
}
\description{
Creates vertical and horizontal grid lines for the psychrometric chart
}
