% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{parse_datetime_column}
\alias{parse_datetime_column}
\title{Parse DateTime Column}
\usage{
parse_datetime_column(datetime_vector)
}
\arguments{
\item{datetime_vector}{Vector of datetime strings}
}
\value{
POSIXct vector
}
\description{
Internal function to parse datetime columns with multiple possible formats
}
\keyword{internal}
