% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{roundmax}
\alias{roundmax}
\title{Round Maximum Value for Color Scale}
\usage{
roundmax(x)
}
\arguments{
\item{x}{Maximum value to round}
}
\value{
Rounded value for color scale
}
\description{
Round Maximum Value for Color Scale
}
\keyword{internal}
