% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric_calculations.R
\name{seasonal_parameters}
\alias{seasonal_parameters}
\title{Calculate Seasonal Parameters}
\usage{
seasonal_parameters(T, H)
}
\arguments{
\item{T}{Temperature vector}

\item{H}{Humidity vector}
}
\value{
Vector of mean temperature, humidity, enthalpy, and specific humidity
}
\description{
Calculate Seasonal Parameters
}
