% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric_calculations.R
\name{sh}
\alias{sh}
\title{Calculate Specific Humidity}
\usage{
sh(H, T, Pressure = 1013.25)
}
\arguments{
\item{H}{Relative humidity percentage}

\item{T}{Temperature in Celsius}

\item{Pressure}{Atmospheric pressure in hPa (default: 1013.25)}
}
\value{
Specific humidity value
}
\description{
Calculate Specific Humidity
}
