% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{try_get_nc_var}
\alias{try_get_nc_var}
\title{Try to Get NetCDF Variable}
\usage{
try_get_nc_var(nc_data, var_names)
}
\arguments{
\item{nc_data}{NetCDF file object}

\item{var_names}{Vector of possible variable names}
}
\value{
Variable data or NULL if not found
}
\description{
Internal helper function to safely extract variables from NetCDF files
}
\keyword{internal}
