% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate Input Parameters}
\usage{
validate_inputs(interval, data_df, animate)
}
\arguments{
\item{interval}{Time interval}

\item{data_df}{Data frame}

\item{animate}{Whether animation is requested}
}
\description{
Internal function to validate input parameters
}
\keyword{internal}
