% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric_calculations.R
\name{wh}
\alias{wh}
\title{Calculate Wet Bulb Humidity}
\usage{
wh(Tl, h)
}
\arguments{
\item{Tl}{Temperature in Celsius}

\item{h}{Enthalpy}
}
\value{
List containing specific humidity and temperature
}
\description{
Calculate Wet Bulb Humidity
}
